@@ -1,87 +1,87 @@
 {
-  "name": "cli31",
-  "version": "0.0.0",
+  "name": "NursingIT.Frontend121",
+  "version": "1.2.1",
   "license": "MIT",
-  "angular-cli": {},
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.4.10",
-    "@angular/compiler": "2.4.10",
-    "@angular/core": "2.4.10",
-    "@angular/forms": "2.4.10",
-    "@angular/http": "2.4.10",
-    "@angular/platform-browser": "2.4.10",
-    "@angular/platform-browser-dynamic": "2.4.10",
-    "@angular/router": "3.4.10",
-    "@types/bootstrap-datepicker": "0.0.5",
+    "@angular/animations": "^4.0.0",
+    "@angular/common": "^4.0.0",
+    "@angular/compiler": "^4.0.0",
+    "@angular/core": "^4.0.0",
+    "@angular/forms": "^4.0.0",
+    "@angular/http": "^4.0.0",
+    "@angular/platform-browser": "^4.0.0",
+    "@angular/platform-browser-dynamic": "^4.0.0",
+    "@angular/router": "^4.0.0",
+    "@types/lodash": "4.14.71",
+    "@types/moment-timezone": "^0.2.32",
+    "@types/bootstrap": "3.3.35",
+    "@types/bootstrap-datepicker": "0.0.9",
     "@types/crypto-js": "^3.1.32",
-    "@types/jquery": "2.0.48",
-    "@types/jwt-decode": "1.4.29",
-    "@types/lodash": "4.14.68",
-    "@types/moment-timezone": "0.2.34",
+    "@types/x-editable": "1.5.28",
+    "@types/typeahead": "0.11.29",
+    "@types/jquery": "3.2.9",
     "angulartics2": "1.1.12",
     "bootstrap": "3.3.7",
-    "bootstrap-datepicker": "1.6.1",
-    "bootstrap-select": "1.11.2",
-    "core-js": "2.4.1",
+    "bootstrap-datepicker": "1.7.1",
+    "bootstrap-select": "1.12.4",
+    "core-js": "^2.4.1",
     "crypto-js": "3.1.7",
-    "datatables": "1.10.12",
+    "datatables": "1.10.13",
     "datatables-responsive": "1.0.7",
-    "jquery": "3.1.1",
+    "jquery": "3.2.1",
     "jwt-decode": "2.1.0",
     "lodash": "4.16.6",
     "moment": "2.15.1",
     "moment-timezone": "0.5.6",
-    "ng2-lazyload-image": "2.2.2",
-    "ng2-page-scroll": "3.1.4",
-    "rxjs": "5.4.2",
+    "ng2-page-scroll": "4.0.0-beta.9",
+    "rxjs": "^5.1.0",
     "tablesaw": "2.0.3",
     "timepicker": "1.11.6",
-    "ts-helpers": "1.1.2",
+    "ts-helpers": "^1.1.1",
     "typeahead.js": "0.11.1",
     "x-editable": "1.5.1",
-    "zone.js": "0.7.8"
+    "zone.js": "^0.8.16"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0-beta.31",
-    "@angular/compiler-cli": "2.4.10",
-    "@types/bootstrap": "3.3.34",
-    "@types/bootstrap-datepicker": "0.0.5",
-    "@types/crypto-js": "3.1.33",
-    "@types/x-editable": "^1.5.27",
-    "@types/typeahead": "0.11.27",
-    "jquery": "3.2.1",
+    "@angular/cli": "1.2.6",
+    "@angular/compiler-cli": "^4.0.0",
+    "@angular/language-service": "^4.0.0",
+    "@types/jasmine": "~2.5.53",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
+    "codelyzer": "~3.0.1",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.1.2",
+    "ts-node": "~3.0.4",
+    "tslint": "~5.3.2",
+    "typescript": "~2.3.3",
     "bower": "1.7.9",
-    "browser-sync": "2.18.12",
-    "codelyzer": "2.0.1",
-    "connect-history-api-fallback": "1.3.0",
+    "browser-sync": "^2.12.8",
+    "connect-history-api-fallback": "^1.2.0",
     "gulp": "3.9.1",
     "gulp-load-plugins": "1.5.0",
     "gulp-print": "2.0.1",
     "gulp-rename": "1.2.2",
     "gulp-shell": "0.6.3",
     "gulp-zip": "4.0.0",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-jasmine": "1.1.0",
-    "karma-coverage-istanbul-reporter": "0.2.3",
     "open": "0.0.5",
-    "protractor": "5.1.2",
     "require-dir": "0.3.2",
-    "run-sequence": "2.0.0",
-    "ts-node": "1.2.1",
-    "tslint": "4.5.1",
-    "typescript": "2.0.10",
-    "webdriver-manager": "10.2.5"
+    "run-sequence": "2.0.0"
   }
 }
