@@ -104,6 +104,7 @@ export class ChooseTestComponent implements OnInit, OnChanges, OnDestroy {
     }
 
     ngOnDestroy(): void {
+        debugger;
         if (this.testsTable)
             this.testsTable.destroy();
         this.tests = [];
@@ -132,14 +133,14 @@ export class ChooseTestComponent implements OnInit, OnChanges, OnDestroy {
         //         this.destinationRoute = event.urlAfterRedirects;
         //     });
         
-        this.router
-            .events
-            .filter(event => event instanceof NavigationStart)
-            .subscribe(e => {
-                this.destinationRoute = e.url;
-            });
+        // this.router
+        //     .events
+        //     .filter(event => event instanceof NavigationStart)
+        //     .subscribe(e => {
+        //         this.destinationRoute = e.url;
+        //     });
 
-
+        debugger;
         this.sStorage = this.common.getStorage();
         if (!this.auth.isAuth())
             this.router.navigate(['/']);
@@ -176,7 +177,9 @@ export class ChooseTestComponent implements OnInit, OnChanges, OnDestroy {
     }
 
 
-    canDeactivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | boolean {
+    canDeactivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot, nextState: RouterStateSnapshot): Observable<boolean> | boolean {
+        debugger;
+        this.destinationRoute = nextState.url;
         let outOfTestScheduling: boolean = this.testService.outOfTestScheduling((this.common.removeWhitespace(this.destinationRoute)));
         if (!this.overrideRouteCheck) {
             if (outOfTestScheduling) {
