@@ -19,12 +19,12 @@ const testsRoutes: Routes = [
     { path: 'tests/add-students', component: AddStudentsComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/review', component: ReviewTestComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/confirmation', component: ConfirmationComponent },
-    { path: 'tests/view/:id', component: ViewTestComponent },
+    { path: 'tests/view/:id', component: ViewTestComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/:action/choose-test/:institutionId', component: ChooseTestComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/:action/schedule-test', component: ScheduleTestComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/:action/add-students', component: AddStudentsComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/:action/review', component: ReviewTestComponent, canDeactivate: [SharedDeactivateGuard] },
-    { path: 'tests/:action/view/:id', component: ViewTestComponent },
+    { path: 'tests/:action/view/:id', component: ViewTestComponent, canDeactivate: [SharedDeactivateGuard] },
     { path: 'tests/:action/confirmation', component: ConfirmationComponent },
     { path: 'tests/testing-session-expired', component: LastTestingSessionComponent },
     { path: 'tests/confirmation-modify-in-progress', component: ConfirmationModifyInProgressComponent }
