@@ -7,13 +7,13 @@ import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/do';
 
 export interface ComponentCanDeactivate {
-  canDeactivate: (route: ActivatedRouteSnapshot, state: RouterStateSnapshot) => boolean | Observable<boolean>;
+  canDeactivate: (route: ActivatedRouteSnapshot, state: RouterStateSnapshot, nextState?:RouterStateSnapshot) => boolean | Observable<boolean>;
 }
 
 @Injectable()
 export class SharedDeactivateGuard implements CanDeactivate<ComponentCanDeactivate> {
-  canDeactivate(component: ComponentCanDeactivate, route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | boolean {
-    return component.canDeactivate ? component.canDeactivate(route, state) : true;
+  canDeactivate(component: ComponentCanDeactivate, route: ActivatedRouteSnapshot, state: RouterStateSnapshot, nextState?: RouterStateSnapshot): Observable<boolean> | boolean {
+    return component.canDeactivate ? component.canDeactivate(route, state, nextState) : true;
   }
 }
 
