@@ -1,7 +1,13 @@
 // Typings reference file, you can add your own global typings here
 // https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
 
-declare var System: any;
+/* SystemJS module definition */
+declare var module: NodeModule;
+interface NodeModule {
+    id: string;
+}
+
+// declare var System: any;
 declare var moment: any;
 declare var _: any;
 declare var CryptoJS :any;
